+++ packages/react/components/dropdown/group/DropdownGroupProps.ts
@@ -1,0 +1,15 @@
+import { View } from 'react-native'
+import { Accessibility } from '../../../objects/facets/Accessibility'
+import { CommonProps } from '../../../objects/facets/CommonProps'
+import { Dev } from '../../../objects/facets/Dev'
+
+/**
+ * DropdownGroup Interface
+ */
+export interface DropdownGroupProps extends Accessibility, Dev, CommonProps {
+  children?: React.ReactNode
+  hideSeparator?: boolean
+}
+
+export type DropdownGroupRef = HTMLDivElement
+export type DropdownGroupNativeRef = View