+++ packages/react/components/dropdown/trigger/index.ts
@@ -1,0 +1,2 @@
+export { default as DropdownTrigger } from './DropdownTrigger'
+export * from './DropdownTriggerProps'