+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Fix calendar order days
 - Fix position on range calendar (web)
 - Fix radio checkbox stretched (web)
+- Update skeleton (native)
+- Dropdown Menu (web)
 - Update and fix skeleton (native)
 - Fix flexbox gap (mobile container query)